moved FootballDataAPITest, deleted unused FootballStatsPredictionsApplicationTests
