--- football-stats-predictions/src/test/kotlin/com/footballdata/football_stats_predictions/FootballStatsPredictionsApplicationTests.kt
@@ -1,12 +1,0 @@
-package com.footballdata.football_stats_predictions
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class FootballStatsPredictionsApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-}
