+++ football-stats-predictions/src/test/kotlin/com/footballdata/football_stats_predictions/data/FootballDataAPITest.kt
@@ -1,6 +1,5 @@
-package com.footballdata.football_stats_predictions
+package com.footballdata.football_stats_predictions.data
 
-import com.footballdata.football_stats_predictions.data.FootballDataAPI
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
